pyproject.toml
[tool.poetry]
name = "terminal-radio"
version = "0.1.0"
description = "A terminal-based internet radio player"
authors = ["Vladislav Boiko <vl4gh7@gmail.com>"]
packages = [{include = "terminal_radio"}]

[tool.poetry.dependencies]
python = "^3.10"
textual = "^0.52.1"
python-vlc = "^3.0.20123"

[tool.poetry.scripts]
radio = "terminal_radio.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
